Question 1
Problem Statement
Design and implement a data structure for a Least Recently Used (LRU) cache. It should support the following operations: get and put.
get(key): Get the value (will always be positive) of the key if the key exists in the cache, otherwise return -1.
put(key, value): Set or insert the value if the key is not already present. When the cache reaches its capacity, it should invalidate the least recently used item before inserting a new item.
Constraints
The number of get and put operations will be in the range [1, 10^5].
The capacity of the cache is between 1 and 10^5.